moved some more stuff from UserPage (old UserModal) into userqueries
