+++ client/src/pages/lab_management/users/AccessCheckCard.tsx
@@ -1,9 +1,8 @@
 import { useEffect, useState } from "react";
 import { Button, Card, CardActions  } from "@mui/material";
-import { AccessCheckExtraInfo } from "./UserPage";
 import { gql, useMutation } from "@apollo/client";
 import AuditLogEntity from "../audit_logs/AuditLogEntity";
-import { GET_USER } from "../../../queries/userQueries";
+import { GET_USER, AccessCheckExtraInfo} from "../../../queries/userQueries";
 
 const APPROVE_CHECK = gql`
   mutation ApproveAccessCheck($id: ID!) {
