+++ client/src/pages/lab_management/users/RestrictionCard.tsx
@@ -1,11 +1,10 @@
 import gql from "graphql-tag";
 import { useCurrentUser } from "../../../common/CurrentUserProvider";
-import { Restriction } from "./UserPage";
 import { Button, Card, CardActions, CardContent, CardHeader, Stack, Typography, useTheme } from "@mui/material";
 import { useMutation } from "@apollo/client";
 import { format, parseISO } from "date-fns";
 import { isStaffFor } from "../../../common/PrivilegeUtils";
-import { GET_USER } from "../../../queries/userQueries";
+import { GET_USER, Restriction } from "../../../queries/userQueries";
 
 interface RestrictionCardProps {
     restriction: Restriction;
